Move the load data code into its own file in lib
This makes the main generate_output script cleaner and has the side-effect of reducing memory usage by allowing irrelevant activation and survey data to be deallocated.
