+++ src/__init__.py
@@ -1,7 +1,7 @@
 from .analysis_file import AnalysisFile
 from .apply_manual_codes import ApplyManualCodes
 from .auto_code import AutoCode
-from .combine_raw_datasets import CombineRawDatasets
+from .load_data import LoadData
 from .production_file import ProductionFile
 from .translate_rapid_pro_keys import TranslateRapidProKeys
 from .ws_correction import WSCorrection
